+++ src/main/java/com/back/domain/schedule/schedule/service/ScheduleService.java
@@ -62,7 +62,7 @@ public class ScheduleService {
     }
 
     @Transactional(readOnly = true)
-    public int countClubSchedules(Long clubId) {
+    public long countClubSchedules(Long clubId) {
         return scheduleRepository.countByClubId(clubId);
     }
